client_d/requirements.txt
grpcio==1.42.0
matplotlib==3.6.2
numpy==1.24.3
Pillow==10.1.0
protobuf==3.20.3
torch==2.0.0
torchvision==0.15.0